requirements.txt
Django==1.8.4
django-ckeditor==5.1.1
